rename doc_query tool
