+++ open_notebook/graphs/doc_query.py
@@ -3,16 +3,16 @@ import os
 from langchain_core.runnables import (
     RunnableConfig,
 )
-from langchain_openai import ChatOpenAI
 from langgraph.graph import END, START, StateGraph
 from loguru import logger
 from typing_extensions import TypedDict
 
 from open_notebook.domain import Note, Notebook, Source
+from open_notebook.model_configs import get_langchain_model
 from open_notebook.prompter import Prompter
 
 
-class AskState(TypedDict):
+class DocQueryState(TypedDict):
     doc_id: str
     doc_content: str
     question: str
@@ -20,11 +20,13 @@ class AskState(TypedDict):
     notebook: Notebook
 
 
-def call_model_with_messages(state: AskState, config: RunnableConfig) -> dict:
-    model = ChatOpenAI(
-        model=os.environ.get("RETRIEVAL_MODEL", os.environ["DEFAULT_MODEL"]),
-        temperature=0,
-    )
+def call_model_with_messages(state: DocQueryState, config: RunnableConfig) -> dict:
+    if config.get("configurable", {}).get("model_name", None):
+        model_name = config.get("configurable", {}).get("model_name", None)
+    else:
+        model_name = os.environ.get("RETRIEVAL_MODEL", os.environ["DEFAULT_MODEL"])
+
+    model = get_langchain_model(model_name)
     system_prompt = Prompter(prompt_template="ask_content").render(data=state)
     logger.debug(f"System prompt: {system_prompt}")
     ai_message = model.invoke(system_prompt)
@@ -32,7 +34,7 @@ def call_model_with_messages(state: AskState, config: RunnableConfig) -> dict:
 
 
 # todo: there is probably a better way to do this and avoid repetition
-def get_content(state: AskState) -> dict:
+def get_content(state: DocQueryState) -> dict:
     doc_id = state["doc_id"]
     if "note:" in doc_id:
         doc: Note = Note.get(id=doc_id)
@@ -42,7 +44,7 @@ def get_content(state: AskState) -> dict:
     return {"doc_content": doc_content}
 
 
-agent_state = StateGraph(AskState)
+agent_state = StateGraph(DocQueryState)
 agent_state.add_node("get_content", get_content)
 agent_state.add_node("agent", call_model_with_messages)
 agent_state.add_edge(START, "get_content")
